Aislando BottomPaginationContent en componentes Creado el BarCodeForm para los codigos de barra
