--- components/InventoryOptions.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const InventoryOptions = () => {
-  return <div>InventoryOptions</div>;
-};
-
-export default InventoryOptions;
