+++ components/tables/TableCustomer.js
@@ -12,7 +12,6 @@ import {
   Dropdown,
   DropdownMenu,
   DropdownItem,
-  User,
   Pagination,
   Tooltip,
 } from "@nextui-org/react";
@@ -22,6 +21,7 @@ import { columnsCustomer } from "@/resources/columnTables";
 import { capitalize } from "@/utils/utils";
 import { removeAccents } from "@/utils/normalized";
 import ModalCustomers from "../modals/ModalCustomers";
+import BottomPaginationContent from "../BottomPaginationContent";
 
 const INITIAL_VISIBLE_COLUMNS = [
   "actions",
@@ -262,45 +262,17 @@ export default function TableCustomer({ customers, deleteCustomer }) {
     hasSearchFilter,
   ]);
 
-  const bottomContent = useMemo(() => {
-    return (
-      <div className="py-2 px-2 flex justify-between items-center">
-        <Pagination
-          isCompact
-          showControls
-          showShadow
-          color="primary"
-          page={page}
-          total={pages}
-          onChange={setPage}
-        />
-        <div className="hidden sm:flex w-[30%] justify-end gap-2">
-          <Button
-            isDisabled={pages === 1}
-            size="sm"
-            variant="flat"
-            onPress={onPreviousPage}
-          >
-            Anterior
-          </Button>
-          <Button
-            isDisabled={pages === 1}
-            size="sm"
-            variant="flat"
-            onPress={onNextPage}
-          >
-            Siguiente
-          </Button>
-        </div>
-      </div>
-    );
-  }, [selectedKeys, items.length, page, pages, hasSearchFilter]);
-
   return (
     <Table
       aria-label="Esto es una tabla de productos"
       isHeaderSticky
-      bottomContent={bottomContent}
+      bottomContent={
+        items.length > 0 ? (
+          <BottomPaginationContent
+            {...{ page, pages, setPage, onNextPage, onPreviousPage }}
+          />
+        ) : null
+      }
       bottomContentPlacement="outside"
       classNames={{
         wrapper: "-z-1 sm:h-[calc(100vh-300px)] scroll",
